Change syscall usec timeouts to unsigned longs
I noticed this due to some sys_block calls having a 'negative' argument in
the printout (due to the %d in the saved string).

While I was here, I also changed halt_core, though note that that timeout
was more of a 'future plan', I think.  The code doesn't use it.

+++ user/parlib/include/parlib/parlib.h
@@ -43,9 +43,9 @@ int			sys_provision(int pid, unsigned int res_type, long res_val);
 int         sys_notify(int pid, unsigned int ev_type, struct event_msg *u_msg);
 int         sys_self_notify(uint32_t vcoreid, unsigned int ev_type,
                             struct event_msg *u_msg, bool priv);
-int         sys_halt_core(unsigned int usec);
+int         sys_halt_core(unsigned long usec);
 void*		sys_init_arsc();
-int         sys_block(unsigned int usec);
+int         sys_block(unsigned long usec);
 int         sys_change_vcore(uint32_t vcoreid, bool enable_my_notif);
 int         sys_change_to_m(void);
 int         sys_poke_ksched(int pid, unsigned int res_type);
